corrections au niveau des formations

+++ utils/deleteImages.js
@@ -1,9 +1,11 @@
-const fs = require('fs');
+const fs = require("fs");
 const deleteFile = (filePath) => {
-      fs.unlink(filePath, (err)=>{
-            if(err){
-                  throw(err)
-            }
-      });
-}
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        throw err;
+      }
+    });
+  }
+};
 exports.deleteFile = deleteFile;
